+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/locate_relay/003_Locate_World.lua
@@ -1,0 +1,16 @@
+--[[mudlet
+type: alias
+name: Locate World
+hierarchy:
+- Levi_Ataxia
+- General
+- Locate Relay
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^lw$
+command: ''
+packageName: ''
+]]--
+
+LocateWorld.start()