+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/locate_relay/004_Locate_Data.lua
@@ -1,0 +1,16 @@
+--[[mudlet
+type: alias
+name: Locate Data
+hierarchy:
+- Levi_Ataxia
+- General
+- Locate Relay
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^locate data (.+)$
+command: ''
+packageName: ''
+]]--
+
+LocateWorld.lookupLocation(matches[2])