+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/locate_relay/005_Locate_Summary.lua
@@ -1,0 +1,16 @@
+--[[mudlet
+type: alias
+name: Locate Summary
+hierarchy:
+- Levi_Ataxia
+- General
+- Locate Relay
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^locate summary$
+command: ''
+packageName: ''
+]]--
+
+LocateWorld.summary()